% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_tsma.R
\name{my_tsma}
\alias{my_tsma}
\title{To construct a moving-average (MA\link{w}) approximation to a time series}
\usage{
my_tsma(Y, w = 5)
}
\description{
To construct a moving-average (MA\link{w}) approximation to a time series
}
