% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myrgb.R
\name{myrgb}
\alias{myrgb}
\title{construct the RGB color histograms for an image via magick(...) functions}
\usage{
myrgb(image, show = FALSE, trim = FALSE)
}
\description{
construct the RGB color histograms for an image via magick(...) functions
}
