% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{myttest}
\alias{myttest}
\title{To conduct a parametric (normality assumed) test on the equality of means for two independent numerical groups.}
\usage{
myttest(x, y, alpha = 0.05, display = FALSE)
}
\description{
To conduct a parametric (normality assumed) test on the equality of means for two independent numerical groups.
}
