% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newrap.R
\name{newrap}
\alias{newrap}
\title{approximate the solution for equations of the form f(x) = 0 via the Newton-Raphson iteration method}
\usage{
newrap(f, x0, tol = 1e-06, maxiter = 250)
}
\description{
approximate the solution for equations of the form f(x) = 0 via the Newton-Raphson iteration method
}
